% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-Prediction.R
\name{load_SLOSModel}
\alias{load_SLOSModel}
\title{Load the SLOS model}
\usage{
load_SLOSModel()
}
\value{
The SLOS model
}
\description{
This function loads the pre-trained model from the package.It's available on GitHub
}
