% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampledData.R
\docType{data}
\name{sampled_data}
\alias{sampled_data}
\title{Sampled Data}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(SampledData)
}
\description{
An anonymized dataset with 1000 entries used for testing the SLOS prediction model.
}
\keyword{datasets}
