% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-Prediction.R
\name{slos_load_pretrained_model}
\alias{slos_load_pretrained_model}
\title{Load the SLOS model}
\usage{
slos_load_pretrained_model()
}
\value{
The SLOS model
}
\description{
This function loads the pre-trained model from the package.It's available on GitHub
}
